requirements.txt
Django==1.8.6
djangorestframework==3.3.1
drf-nested-routers==0.11.1
wheel==0.24.0
